Http(s) compressed resolver appends "?tf-hub-format=compressed" before resolving.
This allows to support both simple file servers which should ignore the query part of
the url (similar to git smart http protocol) and support smart servers which reply with different content if this is a module download by the compressed resolver or is a user
visit of the module.

Remove old "http header" extension based solution which was not as flexible.

PiperOrigin-RevId: 189898926