Add hub2graph script
The problem which this script solves is to convert a Hub module
to the self-contained GraphDef protobuf and also optionally to
Tensorboard logs. E.g.
```
python3 hub2graph.py -m https://tfhub.dev/google/imagenet/mobilenet_v2_140_224/classification/1 -o mobilenetv2.pb -t tblogs
```
Thus it becomes possible to fetch a Hub module and run inference in Go for example, which
does not have the Hub support yet. Besides, it is handly for quick visualization of a module.

Signed-off-by: Vadim Markovtsev <vadim@sourced.tech>