Draw text under top border if object box takes entire image
Sometimes the object box top is at the top of the image.
At the same time, the box bottom can be at the bottom of the image.
In this case, the text is drawn beyond the image.
It this case it's better to draw text right under the top border.