Add hub.text_embedding_column_v2 to work with TF2 SavedModels.
This change is very restrictive in scope to not overdeliver unwanted features before fully testing out those that have been requested.

The scope is made smaller by:
 - adding only dense text feature column,
 - testing only against the following: direct use of DenseFeatures, use inside Keras functional model, use with DnnClassifier, use with DnnEstimator and a dataset.
 - a completely new symbol "text_embedding_column_v2" is created and attempt is not done to create one "smart" symbol that works everywhere.

As usual, this relies on the implementation of feature_columns_v2.StateManagerImpl and some special tricks (but using public APIs) had to be made to get the output shape.

Potential issues:
 - not clear how this behaves regarding distribution strategies, or distributed training in estimator.

PiperOrigin-RevId: 301581442