Fix issue when a mixed precision model with a KerasLayer is saved.
The issue occurred when a mixed precision model uses a KerasLayer wrapping a float32 SavedModel. If the mixed precision model is then saved with an optimizer with slot variables, an error would occur. The issue was that KerasLayer would wrap the SavedModel's variables with AutoCastVariables, which would put both the variables and the AutoCastVariables wrapping them in the checkpoint. This causes issues when there are slot variables.

To fix, KerasLayer no longer creates AutoCastVariables.

PiperOrigin-RevId: 384744914